+++ server/controllers/wazuh-api.ts
@@ -18,6 +18,7 @@ import { KeyEquivalence } from '../../common/csv-key-equivalence';
 import { ApiErrorEquivalence } from '../lib/api-errors-equivalence';
 import apiRequestList from '../../common/api-info/endpoints';
 import { HTTP_STATUS_CODES } from '../../common/constants';
+import { getCustomizationSetting } from '../../common/services/settings';
 import { addJobToQueue } from '../start/queue';
 import fs from 'fs';
 import { ManageHosts } from '../lib/manage-hosts';
@@ -1101,21 +1102,21 @@ export class WazuhApiCtrl {
    * @param request
    * @param response
    */
-  async getAppLogos(context: RequestHandlerContext, request: KibanaRequest, response: KibanaResponseFactory) {
+  async getAppLogos(context: RequestHandlerContext, request: OpenSearchDashboardsRequest, response: OpenSearchDashboardsResponseFactory) {
     try {
       const configuration = getConfiguration();
       const SIDEBAR_LOGO = 'customization.logo.sidebar';
       const APP_LOGO = 'customization.logo.app';
       const HEALTHCHECK_LOGO = 'customization.logo.healthcheck';
 
+      const logos= {
+        [SIDEBAR_LOGO]: getCustomizationSetting(configuration, SIDEBAR_LOGO),
+        [APP_LOGO]: getCustomizationSetting(configuration, APP_LOGO),
+        [HEALTHCHECK_LOGO]: getCustomizationSetting(configuration, HEALTHCHECK_LOGO),
+      }
+
       return response.ok({
-        body: {
-          logos: {
-            [SIDEBAR_LOGO]: configuration[SIDEBAR_LOGO],
-            [APP_LOGO]: configuration[APP_LOGO],
-            [HEALTHCHECK_LOGO]: configuration[HEALTHCHECK_LOGO],
-          }
-        }
+        body: { logos }
       });
     } catch (error) {
       log('wazuh-api:getAppLogos', error.message || error);