+++ server/lib/reporting/summary-table.ts
@@ -10,7 +10,7 @@
  * Find more information about this on the LICENSE file.
  */
 import { Base } from './base-query';
-import { WAZUH_ALERTS_PATTERN } from '../../../common/constants';
+import { getSettingDefaultValue } from '../../../common/services/settings';
 
 interface SummarySetup {
   title: string;
@@ -24,7 +24,7 @@ export default class SummaryTable {
     lte,
     filters,
     summarySetup: SummarySetup,
-    pattern = WAZUH_ALERTS_PATTERN
+    pattern = getSettingDefaultValue('pattern')
   ) {
 
     this._context = context;
@@ -134,7 +134,7 @@ export default class SummaryTable {
  */
   async fetch() {
     try {
-      const response = await this._context.core.elasticsearch.client.asCurrentUser.search({
+      const response = await this._context.core.opensearch.client.asCurrentUser.search({
         index: this._pattern,
         body: this._base
       });