+++ server/routes/wazuh-utils/wazuh-utils.ts
@@ -12,6 +12,7 @@
 import { WazuhUtilsCtrl } from '../../controllers';
 import { IRouter } from 'opensearch_dashboards/server';
 import { schema } from '@osd/config-schema';
+import { CUSTOMIZATION_ENDPOINT_PAYLOAD_UPLOAD_CUSTOM_FILE_MAXIMUM_BYTES, EpluginSettingType, PLUGIN_SETTINGS } from '../../../common/constants';
 
 export function WazuhUtilsRoutes(router: IRouter) {
   const ctrl = new WazuhUtilsCtrl();
@@ -30,13 +31,66 @@ export function WazuhUtilsRoutes(router: IRouter) {
     {
       path: '/utils/configuration',
       validate: {
-        body: schema.object({
-          key: schema.string(),
-          value: schema.any()
-        })
+        body: schema.object(
+          Object.entries(PLUGIN_SETTINGS)
+            .filter(([, { isConfigurableFromFile }]) => isConfigurableFromFile)
+            .reduce(
+              (accum, [pluginSettingKey, pluginSettingConfiguration]) => ({
+                ...accum,
+                [pluginSettingKey]: schema.maybe(
+                  pluginSettingConfiguration.validateBackend
+                    ? pluginSettingConfiguration.validateBackend(schema)
+                    : schema.any()
+                ),
+              }),
+              {}
+            )
+        )
       }
     },
     async (context, request, response) => ctrl.updateConfigurationFile(context, request, response)
+  );
+
+  const pluginSettingsTypeFilepicker = Object.entries(PLUGIN_SETTINGS)
+    .filter(([_, { type, isConfigurableFromFile }]) => type === EpluginSettingType.filepicker && isConfigurableFromFile);
+
+  const schemaPluginSettingsTypeFilepicker = schema.oneOf(pluginSettingsTypeFilepicker.map(([pluginSettingKey]) => schema.literal(pluginSettingKey)));
+
+  // Upload an asset
+  router.put(
+    {
+      path: '/utils/configuration/files/{key}',
+      validate: {
+        params: schema.object({
+          // key parameter should be a plugin setting of `filepicker` type
+          key: schemaPluginSettingsTypeFilepicker
+        }),
+        body: schema.object({
+          // file: buffer
+          file: schema.buffer(),
+        })
+      },
+      options: {
+        body: {
+          maxBytes: CUSTOMIZATION_ENDPOINT_PAYLOAD_UPLOAD_CUSTOM_FILE_MAXIMUM_BYTES,
+        },
+      }
+    },
+    async (context, request, response) => ctrl.uploadFile(context, request, response)
+  );
+
+  // Remove an asset
+  router.delete(
+    {
+      path: '/utils/configuration/files/{key}',
+      validate: {
+        params: schema.object({
+          // key parameter should be a plugin setting of `filepicker` type
+          key: schemaPluginSettingsTypeFilepicker
+        })
+      }
+    },
+    async (context, request, response) => ctrl.deleteFile(context, request, response)
   );
 
   // Returns Wazuh app logs
@@ -45,6 +99,6 @@ export function WazuhUtilsRoutes(router: IRouter) {
       path: '/utils/logs',
       validate: false
     },
-    async (context, request, response) => ctrl.getAppLogs(context,request, response)
+    async (context, request, response) => ctrl.getAppLogs(context, request, response)
   );
 }